+++ src/main/java/com/agibank/corehub/App.java
@@ -10,7 +10,7 @@ import java.io.IOException;
 public class App extends Application {
     @Override
     public void start(Stage stage) throws IOException {
-        FXMLLoader fxmlLoader = new FXMLLoader(App.class.getResource("/com/agibank/corehub/views/login.fxml"));
+        FXMLLoader fxmlLoader = new FXMLLoader(App.class.getResource("/com/agibank/corehub/views/home/login.fxml"));
         Scene scene = new Scene(fxmlLoader.load(), 412, 915);
         stage.setTitle("CoreHub");
         stage.setScene(scene);
