+++ src/main/java/com/agibank/corehub/controller/home/LoginController.java
@@ -1,6 +1,8 @@
-package com.agibank.corehub.controller;
+package com.agibank.corehub.controller.home;
 
 import com.agibank.corehub.beans.Usuario;
+import com.agibank.corehub.controller.Alerta;
+import com.agibank.corehub.controller.CifradorSenha;
 import com.agibank.corehub.dao.UsuarioDAO;
 import java.io.IOException;
 import java.sql.SQLException;
@@ -49,7 +51,7 @@ public class LoginController {
 
 
     public void navegarHome(ActionEvent actionEvent, Usuario usuario) throws IOException {
-        FXMLLoader loader = new FXMLLoader(getClass().getResource("/com/agibank/corehub/views/home.fxml"));
+        FXMLLoader loader = new FXMLLoader(getClass().getResource("/com/agibank/corehub/views/home/home.fxml"));
         Parent root = loader.load();
 
         HomeController homeController = loader.getController();
