+++ src/main/resources/com/agibank/corehub/views/home/login.fxml
@@ -17,13 +17,13 @@
       minWidth="412.0"
       xmlns="http://javafx.com/javafx/21"
       xmlns:fx="http://javafx.com/fxml/1"
-      fx:controller="com.agibank.corehub.controller.LoginController">
+      fx:controller="com.agibank.corehub.controller.home.LoginController">
    <children>
       <Pane prefHeight="200.0" prefWidth="200.0">
          <children>
             <ImageView fitHeight="154.0" fitWidth="150.0" layoutX="25.0" layoutY="32.0" pickOnBounds="true" preserveRatio="true">
                <image>
-                  <Image url="@../assets/logoAgi.png" />
+                  <Image url="@../../assets/logoAgi.png" />
                </image>
             </ImageView>
             <Label layoutX="31.0" layoutY="126.0" prefHeight="39.0" prefWidth="81.0" text="v4.69.1" textAlignment="CENTER">
